feat: aggiunta la possibilità di inserire un parametro quantità all'aggiunta e alla rimozione di un prodotto (in caso non venga indicata il default=1)

+++ app/Http/Controllers/Api/ServiceController.php
@@ -51,6 +51,18 @@ class ServiceController extends Controller
     {
         if ($request->has('cart') && $request->has('product')) {
             $formData = $request->all();
+
+            if ($request->has('quantity')) {
+                $quantityToAdd = $formData['quantity'];
+                if ($quantityToAdd < 1) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Parametri non validi.',
+                    ]);
+                }
+            } else {
+                $quantityToAdd = 1;
+            }
         } else {
             return response()->json([
                 'success' => false,
@@ -74,13 +86,13 @@ class ServiceController extends Controller
             ]);
         }
 
-        // controllo se è già stato aggiunto un prodotto di tipo $randomProduct
-        if ($cart->products()->wherePivot('product_id', $product->id)->exists()) {
+        // controllo se è già stato aggiunto un prodotto di tipo $product
+        if ($cart->products->contains($product)) {
             // in caso affermativo incremento solo la quantità
-            $cart->products()->wherePivot('product_id', $product->id)->increment('quantity');
+            $cart->products()->wherePivot('product_id', $product->id)->increment('quantity', $quantityToAdd);
         } else {
             // altrimenti aggiungo il record
-            $cart->products()->attach($product);
+            $cart->products()->attach($product, ['quantity' => $quantityToAdd]);
         }
 
         return response()->json([
@@ -93,6 +105,18 @@ class ServiceController extends Controller
     {
         if ($request->has('cart') && $request->has('product')) {
             $formData = $request->all();
+
+            if ($request->has('quantity')) {
+                $quantityToRemove = $formData['quantity'];
+                if ($quantityToRemove < 1) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Parametri non validi.',
+                    ]);
+                }
+            } else {
+                $quantityToRemove = 1;
+            }
         } else {
             return response()->json([
                 'success' => false,
@@ -117,12 +141,20 @@ class ServiceController extends Controller
         }
 
         if ($cart->products->contains($product)) {
-            $cart->products()->detach($product);
-
-            return response()->json([
-                'success' => true,
-                'message' => 'Prodotto aggiunto con successo.',
-            ]);
+            $selectedProduct = $cart->products()->wherePivot('product_id', $product->id)->first();
+            if ($selectedProduct->pivot->quantity > $quantityToRemove) {
+                $selectedProduct->pivot->decrement('quantity', $quantityToRemove);
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Quantita decrementata con successo con successo.',
+                ]);
+            } else {
+                $cart->products()->detach($product);
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Prodotto rimosso con successo.',
+                ]);
+            }
         } else {
             return response()->json([
                 'success' => false,
